scheduler/app/database.py
"""
调度器数据库连接配置（与网关模式相同，独立数据库文件）。
"""
import os
from collections.abc import AsyncGenerator

from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine

load_dotenv()

DATABASE_URL: str = os.getenv(
    "SCHEDULER_DATABASE_URL", "sqlite+aiosqlite:///./scheduler.db"
)

engine = create_async_engine(DATABASE_URL, echo=False)
async_session_factory = async_sessionmaker(engine, expire_on_commit=False)


async def get_db() -> AsyncGenerator[AsyncSession, None]:
    async with async_session_factory() as session:
        yield session
