scheduler/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
